Add magic bitboards

+++ src/chess_core/board.jl
@@ -98,7 +98,7 @@ end
 
 ######################### # Square / bit helpers # ######################### 
 "Map (file, rank) → square index (0..63). file=1→a, rank=1→1."
-square_index(file::Int, rank::Int) = (rank - 1) * 8 + (file - 1)
+square_index(file, rank) = (rank - 1) * 8 + (file - 1)
 
 "Map algebraic notation (e.g. 'e3') → square index (0..63)."
 function square_index(sq::AbstractString)
@@ -109,13 +109,13 @@ function square_index(sq::AbstractString)
 end
 
 "Set bit at square sq."
-setbit(bb::UInt64, sq::Int) = bb | (UInt64(1) << sq)
+setbit(bb, sq) = bb | (UInt64(1) << sq)
 
 "Clear bit at square sq."
-clearbit(bb::UInt64, sq::Int) = bb & ~(UInt64(1) << sq)
+clearbit(bb, sq) = bb & ~(UInt64(1) << sq)
 
 "Check if bit at square sq is set."
-testbit(bb::UInt64, sq::Int) = ((bb >> sq) & 0x1) == 1
+testbit(bb, sq) = ((bb >> sq) & 0x1) == 1
 
 ######################### Zobrist hashing #########################
